+++ test/unittests/test_tokenization.py
@@ -30,7 +30,7 @@ class TestTokenization(unittest.TestCase):
     CONFIG_TYPE = PluginConfigTypes.TOKENIZATION
     TEST_CONFIG = {"test": True}
     CONFIG_SECTION = "tokenization"
-    TEST_LANG = "en-us"
+    TEST_LANG = "en-US"
 
     @patch("ovos_plugin_manager.utils.find_plugins")
     def test_find_plugins(self, find_plugins):